Revert "Use generics instead of traits for mocking"
This reverts commit e2bba69f2c5f10193fe4165b2d115ea6aee1ef11.

Changed mind about the generics approach. It's reasonably tidy now, but
feels like it would quickly become boilerplatey and noisy if the number
of dependencies grew, and liable to run into dead-ends as an approach if
the things being mocked out need to become more dynamic. It seems to me
that the overhead of Box is totally acceptable for this tool in exchange
for more flexible mocking.