Stream stdout/stderror
Spawn the process and then wait for it to finish.

Turns out you don't need to capture stdout at all and it's streamed to
the program's stdout/err by default. Win. You only need to capture
stdout if you want to look at it in rust code, which I don't, it just
needs to be relayed to the user of gitopolis.