Pull in Nathan's OS specific test code
Although we managed to get the windows output to end up the same, it
turns out the OSX output is actually different so we need some kind of
OS specific test switching after all.

Pulling in the patch from Nathan verbatim with a view to morphing it
into an OSX variant.

From:
- https://github.com/rustworkshop/gitopolis/pull/120
- https://github.com/rustworkshop/gitopolis/compare/main...NathanLBCooper:gitopolis:tests-on-windows

Co-authored-by: Nathan Cooper <nathan.cooper@tradera.com>